requirements.txt
requests
google-cloud-bigquery
google.cloud
prettytable
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
db-dtypes
croniter
pytz
